+++ app/uk/gov/hmrc/gform/wshttp/WSHttpModule.scala
@@ -17,7 +17,7 @@
 package uk.gov.hmrc.gform.wshttp
 
 import uk.gov.hmrc.gform.config.ConfigModule
-import uk.gov.hmrc.gform.theauditing.AuditingModule
+import uk.gov.hmrc.gform.auditing.AuditingModule
 import uk.gov.hmrc.play.http.hooks.HttpHook
 import uk.gov.hmrc.play.http.ws.WSHttp
